aggregation/requirements.txt
pandas
boto3
psycopg2-binary
python-dotenv
sqlalchemy
flask-sqlalchemy
s3fs